Abstract MCMC

Types

MCMC.AbstractMCMC — Type

AbstractMCMC

Supertype for all MCMC algorithms. It provides a common interface for sampling, updating, and saving the MCMC state.

All subtypes of AbstractMCMC should implement the following methods:

sample(<:AbstractMCMC): Sample a new point from the proposal distribution.

update!(<:AbstractMCMC, x_test): Update the MCMC state with the new point. It should return

true if the configuration has been modified, and false otherwise.

Optionally subtypes of AbstractMCMC can overload the function should_save to define when the MCMC state should be saved. By default, the state will never be saved. In order to save the state, the method HDF5.write(::HDF5.File, ::String, <:AbstractMCMC) must be defined.

Furthermore the following symbols should be defined for the subtype either as fields or as methods:

rng(<:AbstractMCMC): Return the random number generator to be used for sampling.

observables(<:AbstractMCMC): Return a list of observables to be computed,

observables must be functions that take the mcmc as argument.

save_file(<:AbstractMCMC): Return an IO stream where the the obervables output will be saved.

checkpoint_file(<:AbstractMCMC): Return the filename to save the MCMC state. Alternatively,

a checkpoint_file field can be defined for the subtype.

Methods

MCMC.checkpoint_file — Function

checkpoint_file(mcmc::AbstractMCMC)

If the concrete subtype of AbstractMCMC has checkpoint_file as a field, it will be returned, otherwise a MethodError will be thrown.

MCMC.save_file — Function

save_file(mcmc::AbstractMCMC)

If the concrete subtype of AbstractMCMC has save_file as a field, it will be returned, otherwise a MethodError will be thrown.

MCMC.rng — Function

rng(mcmc::AbstractMCMC)

If the concrete subtype of AbstractMCMC has rng as a field, it will be returned, otherwise a MethodError will be thrown.

MCMC.observables — Function

observables(mcmc::AbstractMCMC)

If the concrete subtype of AbstractMCMC has observables as a field, it will be returned, otherwise a MethodError will be thrown.

MCMC.should_save — Function

should_save(mcmc::AbstractMCMC, i::Int)

Determine if the MCMC state should be saved. By default, it returns false. Subtypes of AbstractMCMC can overload this function to define when the state should be saved. The default implementation is to never save the state.

MCMC.save! — Function

save!(mcmc::AbstractMCMC)

Save the MCMC state to the checkpoint HDF5 file calling HDF5.write.

MCMC.run! — Function

run!(mcmc::AbstractMCMC, n::Int)

Run the MCMC algorithm for n iterations. It will sample a new point, update the configuration, compute the observables and save the state when should_save returns true.

The steps are logged at debug level. For utility you can use the mcmc_logger function to:

with_logger(mcmc_logger()) do
	run!(mcmc, n)
end