BaseMCMC

AbastractBaseMCMC is a slightly more specific abstract type to work with, which can be used for all the concrete types that only require the id, rng, state, checkpoint_file and save_file, where the type of the state field is a parameter of the type.

Types

MCMC.AbstractBaseMCMC — Type

AbstractBaseMCMC{T} <: AbstractMCMC

Base abstract class for MCMC algorithms. The type parameter T is the type of the state of the MCMC. Concrete subtypes are required to have the following fields:

id::UUID: Unique identifier for the MCMC instance.

rng::Xoshiro: Random number generator, default is seeded with the default generator.

state::T: The state of the MCMC, type T is the type of the state.

checkpoint_file::HDF5.File: File for checkpointing, default is created using default_checkpoint_file.

save_file::IO: File for saving the MCMC state, default is created using default_save_file.

Also, the type T must have implemented the method HDF5.write

The default constructor is provided:

BaseMCMC_T(state::T; id::UUID=uuid4(), rng::Xoshiro=Xoshiro([rand(UInt64) for _ in 1:5]...),
		checkpoint_file::HDF5.File=default_checkpoint_file(id), save_file_::IO=default_save_file(id))

Methods

MCMC.state — Function

state(mcmc::AbstractBaseMCMC)

Return the state of the MCMC.

MCMC.rng_state — Function

rng_state(mcmc::AbstractBaseMCMC)

Return the state of the Xoshiro rng used by the MCMC.

Base.write — Method

HDF5.write(parent::Union{HDF5.File,HDF5.Group}, group_name::String, mcmc::AbstractBaseMCMC)

Write the MCMC state to the HDF5 file. The state type must implement the HDF5.write method.

Macros

MCMC.@BaseMCMC_def — Macro

@BaseMCMC_def(T, expr)

Macro to define a keyword constructor for AbstractBaseMCMC{T} concrete subtypes. The symbol T is the type of the state of the MCMC.

If other fields are defined in the struct, they will be preserved. Mind that in this case the keyword constructor will fail, so you will have to define it manually.

@macroexpand @BaseMCMC_def Vector{Float64} struct PiMCMC <: AbstractPiMCMC end

# output

:(struct PiMCMC <: Main.AbstractPiMCMC
      #= /tmp/jl_6fDa5dH.jl:7 =#
      id::Main.UUID
      #= /tmp/jl_6fDa5dH.jl:8 =#
      rng::Main.Xoshiro
      #= /tmp/jl_6fDa5dH.jl:9 =#
      state::Main.Vector{Main.Float64}
      #= /tmp/jl_6fDa5dH.jl:10 =#
      checkpoint_file::(Main.HDF5).File
      #= /tmp/jl_6fDa5dH.jl:11 =#
      save_file::Main.IO
      #= /tmp/jl_6fDa5dH.jl:17 =#
  end)