PiMCMC

Types

MCMC.AbstractPiMCMC — Type

AbstractPiMCMC <: AbstractMCMC

Abstract type for PiMCMC algorithms. Concrete subtypes must have implemented either a method for the state and for the random number generator or have them as fields as state and rng.

MCMC.PiMCMC — Type

PiMCMC <: AbstractPiMCMC <: AbstractMCMC

Concrete type for PiMCMC algorithms.

The rng is a Xoshiro, by default it seeded with the default generator.

The checkpointfile is defaulted using [`defaultcheckpointfile`](@ref defaultcheckpoint_file).

The savefile is defaulted using [`defaultsavefile`](@ref defaultsave_file).

The closing of the file is automatically handled by the finalizer.

The rng_state method is implemented for the type, but not the should_save.

Methods

MCMC.state — Method

state(mcmc::AbstractPiMCMC)

Returns the state of the MCMC as a field.

MCMC.rng — Method

rng(mcmc::AbstractPiMCMC)

Returns the random number generator of the MCMC as a field.

MCMC.sample — Method

sample(mcmc::AbstractPiMCMC)

Sample a new point in the [0, 1]×[0, 1] square.

MCMC.update! — Method

update!(mcmc::AbstractPiMCMC, x_test::Vector{Float64})

Update the state of the MCMC with the new point, return always true.

MCMC.observables — Method

observables(mcmc::AbstractPiMCMC)

Returns the observables of the PiMCMC.

Base.write — Method

HDF5.write(file::HDF5.File, group_name::String, mcmc::AbstractPiMCMC)

Write the state of the MCMC to the HDF5 file. The concrete type must have implemented rng_state

MCMC.rng_state — Method

rng_state(mcmc::PiMCMC)

Returns the state of the Xoshiro rng as a vector.